Re-structure service files based on function
